Conflicts:
	CMakeLists.txt [version]
	cmake/modules/FindKDE4Internal.cmake [moved to ECM, merged there]
	kio/tests/CMakeLists.txt [new tests, moved to tier3/kio/autotests]
	tier1/kconfig/src/core/kcoreconfigskeleton.cpp [if sync() vs return false]
	tier1/kconfig/src/gui/kconfigloader.cpp [git got confused with plasma/configloader.cpp!]
	tier1/kcoreaddons/src/lib/io/kdirwatch.cpp [qDebug vs removeEntry]
	tier1/kcoreaddons/src/lib/io/kdirwatch_p.h [qDebug vs removeEntry]
	tier1/solid/src/solid/backends/udev/udevdevice.cpp [tr usage fix]
	tier1/solid/src/solid/backends/udev/udevprocessor.cpp [moc include]
	tier1/solid/src/solid/backends/win/winopticaldrive.cpp [moc include]
	tier3/kio/src/core/global.h [new func vs moved out]
	tier3/kio/src/core/job_error.cpp [new func vs moved out]
	tier3/kio/src/ioslaves/http/http.cpp [qDebug]
	tier4/khtml/src/khtml_part.cpp [KGuiItem in context]

+++ tier1/kjs/src/kjs/identifier.cpp
@@ -141,7 +141,7 @@ PassRefPtr<UString::Rep> Identifier::add(const UChar *s, int length)
          return &UString::Rep::empty;
     }
 
-    UCharBuffer buf = {s, length};
+    UCharBuffer buf = {s, static_cast<unsigned int>(length)};
     return *identifierTable().add<UCharBuffer, UCharBufferTranslator>(buf).first;
 }
